change the prefix of the second tempfile
There seem to be strange issues popping up with rubinius. Amongst these are that the tempfile is just not created, i.e. the path is empty which crashes tests further down. It might be due to the similarity in this line, though at this point I'm shooting random. I have recreated the failed test exactly the way travis runs them (commandwise) and the problem is not showing up.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>